update: move type
