+++ src/types/index.ts
@@ -8,3 +8,20 @@ export interface RankingResponse {
   rankings?: RankingI[];
   user?: { score: number; ranking: number };
 }
+
+export enum University {
+  Kaist = 'K',
+  Postech = 'P'
+}
+
+export enum Winner {
+  Kaist = 'K',
+  Postech = 'P',
+  None = 'N'
+}
+
+export enum GameStatus {
+  Waiting = 'waiting',
+  Running = 'running',
+  Exiting = 'exiting'
+}
